My CMPXCHG16B idea was not going to work. Used a compressed pointers instead.
The compare exchange method I had though of was either going to have a
race in it or was going to be extremely slow. Therefore I moved to using
a compressed pointer to hold both the top level bucket pointer and log2
of the modulus. Since the modulus is always a power of 2 we can do this
easily. Also we only need to store up to 64 in the upper portion of the
pointer so running out of space is not an issue.